\name{hipr-package}
\alias{hipr-package}
% - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - %
% Alias
% - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - %
\alias{hipr}
\docType{package}
\title{Package which contains (C and R) routines for hip_frac project.}
\description{repTable function which replicates database rows.}
\details{
\tabular{ll}{
Package: \tab hipr\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-12-04\cr
License: \tab What license is it under?\cr
}
}
\author{
Wolfgang Kaisers
Maintainer: Wolfgang Kaisers <kaisers@med.uni-duesseldorf.de>
}
\keyword{ package }
\examples{
#
}
